# openapi.yaml
openapi: 3.0.0
info:
  title: MindPal API
  description: Complete documentation of all API endpoints for the MindPal Mental Health Application.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /serenity/chat:
    post:
      tags: [Serenity AI Chat]
      summary: Send message to Serenity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                  example: How can I improve my sleep quality?
                sessionId:
                  type: string
                  description: Optional. Pass a valid session ID to continue an existing conversation. Leave empty to start a new session.
                  example: ""
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    type: object
                    properties:
                      sessionId:
                        type: string
                      message:
                        type: string
                      sentiment:
                        type: object
                      timestamp:
                        type: string
                        format: date-time
...
  /serenity/session/{id}:
    get:
      tags: [Serenity AI Chat]
      summary: Get specific session
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            description: A valid MongoDB ObjectId for the session
            example: 60d0fe4f5311236168a109ca
      responses:
        "200":
          description: Session details
  /serenity/session:
    post:
      tags: [Serenity AI Chat]
      summary: Create new session
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [text, voice]
                  example: text
      responses:
        "201":
          description: Session created
...
  /serenity/suggestions:
    get:
      tags: [Serenity AI Chat]
      summary: Get activity suggestions
      parameters:
        - in: query
          name: mood
          schema:
            type: integer
            minimum: 1
            maximum: 10
            example: 5
        - in: query
          name: timeOfDay
          schema:
            type: string
            enum: [morning, afternoon, evening, night]
            example: afternoon
      responses:
        "200":
          description: Suggestions list
  /serenity/permissions:
    get:
      tags: [Serenity AI Chat]
      summary: Get data permissions
      responses:
        "200":
          description: User permissions

    post:
      tags: [Serenity AI Chat]
      summary: Update permissions
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                permissions:
                  type: object
                  properties:
                    analyzeMood:
                      type: boolean
                      example: true
                    analyzeSleepData:
                      type: boolean
                      example: true
                    analyzeJournals:
                      type: boolean
                      example: true
      responses:
        "200":
          description: Permissions updated
...
  /community/posts:
    post:
      tags: [Community]
      summary: Create post
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                isAnonymous:
                  type: boolean
                tags:
                  type: array
                  items:
                    type: string
      responses:
        "201":
          description: Post created
  /community/posts/{id}/like:
    post:
      tags: [Community]
      summary: Like/unlike post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Like status updated
  /community/posts/{id}/comments:
    post:
      tags: [Community]
      summary: Add comment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Comment added
    get:
      tags: [Community]
      summary: Get comments
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Comments list

  # Therapists
...